Revert (move Provider plugin again) "Revert "move `generic-provider-views` and `Provider` to /plugins/Provider""
This reverts commit 49b4480222b1361a6a6de86539d4b1def0f07a0c.
