+++ src/plugins/GoogleDrive/index.js
@@ -1,9 +1,7 @@
 const html = require('yo-yo')
 const Plugin = require('../../core/Plugin')
-
-const Provider = require('../../Provider')
-
-const View = require('../../generic-provider-views')
+const Provider = require('../Provider')
+const View = require('../Provider/view')
 
 module.exports = class GoogleDrive extends Plugin {
   constructor (core, opts) {
