+++ src/plugins/Provider/view/Browser.js
